Show executions in detail page
